refactored input type function, rearranged files
