+++ ruby/translate_braille_to_english.rb
@@ -1,0 +1,3 @@
+def translate_braille_to_english(braille_string)
+  result = ""
+end
